Moved original files to directory hierarchy.
Note that we may need to add absolute path names to the .sh
scripts, since they don't live in the same folder as the .R
scripts anymore. We can figure out how to do this with relative
paths. I have a trick for bash to figure out the enclosing
directory that is home to the .sh script and that can work from
there.